Add ability to stream audio to server
Can now stream audio in the form of a wav file to server. Similar to how
video stream is transmitted, the audio is sliced up into a time interval
slice. Future work will require the server to stitch up the audio and
video together.

More work needs to be done on making the bug less error prone. At the
moment the code assumes good connection. But need to add in more robust
code if connection is broken, such as broken websockets.